# 6sem/PIS/Lab_11/GO11_01/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для CRUD операций со знаменитостями (лабораторная 11).",
        "title": "GO11_01 Celebrities API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/Celebrities": {
            "post": {
                "description": "POST /Celebrities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "celebrities"
                ],
                "summary": "Создать знаменитость",
                "parameters": [
                    {
                        "description": "Данные знаменитости",
                        "name": "celebrity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Celebrity"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Celebrity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Celebrities/All": {
            "get": {
                "description": "GET /Celebrities/All",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "celebrities"
                ],
                "summary": "Получить всех знаменитостей",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Celebrity"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Celebrities/{id}": {
            "get": {
                "description": "GET /Celebrities/{id}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "celebrities"
                ],
                "summary": "Получить знаменитость по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID знаменитости",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Celebrity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "PUT /Celebrities/{id}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "celebrities"
                ],
                "summary": "Обновить знаменитость",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID знаменитости",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные знаменитости",
                        "name": "celebrity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Celebrity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Celebrity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "DELETE /Celebrities/{id}",
                "tags": [
                    "celebrities"
                ],
                "summary": "Удалить знаменитость",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID знаменитости",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Celebrity": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string",
                    "example": "Leonardo DiCaprio"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nationality": {
                    "type": "string",
                    "example": "American"
                },
                "reqPhotoPath": {
                    "type": "string",
                    "example": "/images/dicaprio.jpg"
                }
            }
        }
    }
}